# YJIT: Optimize Integer#succ